Moved sync message to earlier moment in time
Sync took place on first call to transmit_message(), which could have introduced race conditions because the message structures where already initialised for sending the packet passed from the userspace to the board controller and we had a recursive call to perform_transaction().
Now the syncing takes place on the first call of transmit_packet from higher level kernel modules.

Signed-off-by: Christina Quast <chrysh.ng+github@gmail.com>