[drbcc-wd] Implemented workqueue for keepalive calls
When timer callback is triggered, it adds a keepalive ping work into a workqueue.
In the work function eventually the heartbeat (ping) message will be sent out by drbcc-core.
The workqueue was needed to be added since the timer callback was issued in interrupt context but the transmit_packet function of the drbcc-core may sleep.

Signed-off-by: Christina Quast <chrysh.ng+github@gmail.com>