[drbcc-pkt] Return -EAGAIN if pkt is partly parsed
In case a buffer is passed to the core module where only the start byte or a start byte and the command are contained, the parsing function should rather return an error value instead of the bytes parsed.
This worked for message parts longer than 2 bytes before.
But the start and command bytes have to be treated differently from data bytes. So there was a bug when parsing these two bytes.

Signed-off-by: Christina Quast <chrysh.ng+github@gmail.com>